Add a test for isolated bundles in monolith output
